oilfield_agent/requirements.txt
langgraph>=0.2.0
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-anthropic>=0.2.0
langchain-community>=0.3.0
sqlalchemy>=2.0.0
pyodbc>=5.0.0
pandas>=2.0.0
numpy>=1.26.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
fastapi>=0.115.0
uvicorn>=0.30.0
streamlit>=1.39.0
faker>=28.0.0
rich>=13.0.0
